# Conflicts:
#	BE/src/main/java/com/ssafy/IBG/InitDB.java


Former-commit-id: 8ccfb9a6ee6333e7c5e834dd501295cc648c3549

+++ BE/src/main/java/com/ssafy/IBG/domain/Score.java
@@ -25,7 +25,7 @@ public class Score {
     @JoinColumn(name = "gameNo")
     private Game game;
 
-    private int scoreRating;
+    private double scoreRating;
 
     public Score(User user, Game game, int scoreRating){
         this.user = user;
